Restructure the whole project
All code was converted to vue.js single file components which combine
the javascript, css, and html that make up a component into a single
file. To accomodate this change, a host of new tools were taken into
use, npm and webpack being the most significant ones. External
dependencies were removed from the repository and are now managed with
npm.