Make the large graph popup responsive
The large popup and its graph have hardcoded wider dimensions based on
a window width breakpoint. The graph won't redraw itself if its or the
target element's size changes which makes this a bit tricky. If the
window's size changes (in practice only on a laptop/desktop computer),
the popup must be reopened for the graph's size to be corrected.