Prevent flashing of popup's dose rate on datetime change
The change event leading to this function triggers generically multiple
times on source changes with no info on why. The source typically has no
features the first time the event triggers which results in the dose
rate flashing first to the dash character and then to the value the next
time the event triggers.