Get rid of relative imports
It's easier to change directory structure with absolute paths as you can
just search and replace the imports. It's also easier to tell where
things come from at a glance instead of having to mentally parse the
relative paths.

The jest config setting is needed to in order to use "@" alias in imports.