Fix getting of large amounts of data with span option
The HTTP responses are actually supposed to be read from right away
and then closed (which the 'with' does). Previously at the end of
getting say 200 datasets the whole thing would fail because the HTTP
responses were kept around until everything was downloaded. Now you
should be able to load large amounts of data with a single command.